refactor(ui): extract message-list viewport rendering
